Chatbot/backend/requirements.txt
fastapi
uvicorn[standard]
openai
langchain
pydantic
sentence-transformers
faiss-cpu
python-dotenv
requests
httpx
typing-extensions
